move API methods to classes named after common endpoint prefix
especially
- splitting Public into Instances and Search
- moving OAuth-related from Apps into OAuth
