+++ sample-java/src/main/java/social/bigbone/sample/PerformSimpleSearch.java
@@ -3,7 +3,7 @@ package social.bigbone.sample;
 import social.bigbone.MastodonClient;
 import social.bigbone.api.entity.Results;
 import social.bigbone.api.exception.BigBoneRequestException;
-import social.bigbone.api.method.Public;
+import social.bigbone.api.method.Search;
 
 @SuppressWarnings("PMD.SystemPrintln")
 public class PerformSimpleSearch {
@@ -18,8 +18,8 @@ public class PerformSimpleSearch {
             .build();
 
         // Perform search and print results
-        final Public publicMethod = new Public(client);
-        final Results searchResult = publicMethod.search(searchTerm).execute();
+        final Search searchMethod = new Search(client);
+        final Results searchResult = searchMethod.searchContent(searchTerm).execute();
         searchResult.getAccounts().forEach(account -> System.out.println(account.getDisplayName()));
         searchResult.getStatuses().forEach(status -> System.out.println(status.getContent()));
         searchResult.getHashtags().forEach(hashtag -> System.out.println(hashtag.getName()));
